pyproject.toml
[tool.poetry]
name = "elderly-backend"
version = "0.1.0"
description = ""
authors = ["fecaridade <caridade500@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
